+++ src/main/java/de/ailis/usb4java/libusb/ControlSetup.java
@@ -37,9 +37,9 @@ public final class ControlSetup
         return controlSetup.get(0);
     }
 
-    public void setBmRequestType(final int bmRequestType)
+    public void setBmRequestType(final byte bmRequestType)
     {
-        controlSetup.put(0, (byte) bmRequestType);
+        controlSetup.put(0, bmRequestType);
     }
 
     public byte bRequest()
@@ -47,9 +47,9 @@ public final class ControlSetup
         return controlSetup.get(1);
     }
 
-    public void setBRequest(final int bRequest)
+    public void setBRequest(final byte bRequest)
     {
-        controlSetup.put(1, (byte) bRequest);
+        controlSetup.put(1, bRequest);
     }
 
     public short wValue()
@@ -57,9 +57,9 @@ public final class ControlSetup
         return controlSetup.getShort(2);
     }
 
-    public void setWValue(final int wValue)
+    public void setWValue(final short wValue)
     {
-        controlSetup.putShort(2, (short) wValue);
+        controlSetup.putShort(2, wValue);
     }
 
     public short wIndex()
@@ -67,9 +67,9 @@ public final class ControlSetup
         return controlSetup.getShort(4);
     }
 
-    public void setWIndex(final int wIndex)
+    public void setWIndex(final short wIndex)
     {
-        controlSetup.putShort(4, (short) wIndex);
+        controlSetup.putShort(4, wIndex);
     }
 
     public short wLength()
@@ -77,8 +77,8 @@ public final class ControlSetup
         return controlSetup.getShort(6);
     }
 
-    public void setWLength(final int wLength)
+    public void setWLength(final short wLength)
     {
-        controlSetup.putShort(6, (short) wLength);
+        controlSetup.putShort(6, wLength);
     }
 }
