+++ src/main/java/de/ailis/usb4java/libusb/Transfer.java
@@ -82,7 +82,7 @@ public final class Transfer
      * @param flags
      *            The transfer flags to set.
      */
-    public native void setFlags(final int flags);
+    public native void setFlags(final byte flags);
 
     /**
      * Returns the address of the endpoint where this transfer will be sent.
@@ -97,7 +97,7 @@ public final class Transfer
      * @param endpoint
      *            The endpoint address to set
      */
-    public native void setEndpoint(final int endpoint);
+    public native void setEndpoint(final byte endpoint);
 
     /**
      * Returns the type of the endpoint.
@@ -112,7 +112,7 @@ public final class Transfer
      * @param type
      *            The endpoint type to set.
      */
-    public native void setType(final int type);
+    public native void setType(final byte type);
 
     /**
      * Returns the timeout for this transfer in milliseconds. A value of 0
